+++ backend/core/src/test/java/site/timecapsulearchive/core/common/fixture/dto/GroupDtoFixture.java
@@ -2,7 +2,7 @@ package site.timecapsulearchive.core.common.fixture.dto;
 
 import java.util.List;
 import site.timecapsulearchive.core.domain.group.data.dto.GroupCreateDto;
-import site.timecapsulearchive.core.domain.group.data.dto.GroupOwnerSummaryDto;
+import site.timecapsulearchive.core.domain.group_member.data.GroupOwnerSummaryDto;
 
 public class GroupDtoFixture {
